Refactor API service in its own function
