+++ server/web/src/components/Pages/GenericQuestion/GenericQuestion.tsx
@@ -15,49 +15,9 @@ import { HourglassLoader } from '@/components/HourglassLoader';
 import { LoadingContext } from '@/state/Loading';
 import { SettingsContext } from '@/state/Settings';
 
-import {
-  ApiOptions,
-  EndpointsEnum,
-  apiConfigConstructor,
-  apiFetch,
-  defaultApiServer,
-} from '@/utils/api';
+import { postChatCompletions } from '@/utils/api';
 
 import styles from './GenericQuestion.module.css';
-
-const baseHeaders = {
-  Accept: 'application/json',
-  'Content-Type': 'application/json',
-};
-
-const chatCompletionsBaseRequest: ChatCompletionsRequest = {
-  model: 'gpt-3.5-turbo-16k',
-  messages: [
-    {
-      role: 'user',
-      content: '',
-      name: 'USER',
-    },
-  ],
-  temperature: 0.4,
-  top_p: 1.0,
-  n: 1,
-  max_tokens: 12847,
-  presence_penalty: 0.0,
-  frequency_penalty: 0.0,
-  logit_bias: {},
-  user: 'USER',
-};
-
-const chatCompletionsApiBaseOptions: ApiOptions = {
-  endpointServer: defaultApiServer,
-  endpointPath: EndpointsEnum.chatCompletions,
-  endpointValue: '',
-  requestOptions: {
-    method: 'POST',
-    headers: baseHeaders,
-  },
-};
 
 export function GenericQuestion() {
   const [loading, setLoading] = useContext(LoadingContext);
@@ -73,34 +33,13 @@ export function GenericQuestion() {
         setLoading(true);
         console.group(`🖱️ Generic question form used:`);
 
-        const chatCompletionsRequest: ChatCompletionsRequest = {
-          ...chatCompletionsBaseRequest,
-          messages: [
-            {
-              ...chatCompletionsBaseRequest.messages[0],
-              content: prompt,
-            },
-          ],
-        };
-        const chatCompletionsApiOptions: ApiOptions = {
-          ...chatCompletionsApiBaseOptions,
-          body: JSON.stringify(chatCompletionsRequest),
-          requestOptions: {
-            ...chatCompletionsApiBaseOptions.requestOptions,
-            headers: {
-              ...chatCompletionsApiBaseOptions.requestOptions?.headers,
-              Authorization: `Bearer ${settings.apiKey}`,
-            },
-          },
-        };
-        const chatCompletionsApiConfig = apiConfigConstructor(
-          chatCompletionsApiOptions,
-        );
-        const chatCompletionResponse = await apiFetch<ChatCompletionsResponse>(
-          chatCompletionsApiConfig,
-        );
-        const { content } = chatCompletionResponse.choices[0].message;
+        if (!settings.apiKey) throw 'API key not set';
 
+        const chatCompletionResponse = await postChatCompletions({
+          apiKey: settings.apiKey,
+          prompt,
+        });
+        const { content } = chatCompletionResponse.choices[0].message;
         setResponseMessage(content);
 
         console.info(`Chat completions request completed`);
@@ -108,7 +47,7 @@ export function GenericQuestion() {
         const userFriendlyError = `Chat completions request couldn't be completed`;
         console.info(userFriendlyError);
         setShowAlert(`
-              ${userFriendlyError}, is the API key set?`);
+              ${userFriendlyError}. ${error}`);
       } finally {
         console.groupEnd();
         setLoading(false);
@@ -190,7 +129,7 @@ export function GenericQuestion() {
       <Snackbar
         open={!!showAlert}
         onClose={(_, reason) => reason !== 'clickaway' && setShowAlert('')}
-        autoHideDuration={3000}>
+        autoHideDuration={5000}>
         <Alert severity="error">{showAlert}</Alert>
       </Snackbar>
     </>
